Add files via upload
A program that accepts user input for quantity sold of 4 unique items and calculates a salesperson's earnings based on a base salary plus commission.  The program allows for data on multiple salespeople to be input and then displays the salary distribution, i.e. number of employees that fall into each of 9 different pay bands.