Add files via upload
A program that creates an empty array of 100 int values, populates the array with a random number generator, and prints the values to a txt file.  It then reads the values back from the text file, stores them back to the array, sorts the array in ascending order, and then prints the values to a new txt file.